requirements.txt
python-dotenv~=1.1.1
requests~=2.32.5
pillow~=12.0.0